complete feature compliance checking and add documentation
