+++ docs/user/index.md
@@ -32,6 +32,7 @@ The repository is also responsible for importing and  exporting CSARs, the excha
 - [Splitting](Splitting) - splitting functionality
 - [TopologyCompletion](TopologyCompletion) - topology completion with a [Tutorial](TopologyCompletionTutorial)
 - [XaaSPackager](XaaSPackager) - Enables reusing modeled topologies as templates for single applications
+- [Compliance Checking](ComplianceChecking.md) - Enables compliance checking of Topology Templates based on reusable Compliance Rules  
 
 ## Background Literature
