+++ org.eclipse.winery.repository.rest/src/main/java/org/eclipse/winery/repository/rest/resources/compliancerules/CRTopologyTemplateResource.java
@@ -33,92 +33,91 @@ import org.slf4j.LoggerFactory;
 
 public class CRTopologyTemplateResource {
 
-	private static final Logger LOGGER = LoggerFactory.getLogger(CRTopologyTemplateResource.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(CRTopologyTemplateResource.class);
 
-	private final TTopologyTemplate topologyTemplate;
+    private final TTopologyTemplate topologyTemplate;
 
-	private final ComplianceRuleResource complianceRuleRes;
+    private final ComplianceRuleResource complianceRuleRes;
 
-	/**
-	 * A topology template is always nested in a service template
-	 */
-	public CRTopologyTemplateResource(ComplianceRuleResource parent, TTopologyTemplate template) {
-		this.topologyTemplate = template;
-		this.complianceRuleRes = parent;
-	}
+    /**
+     * A topology template is always nested in a service template
+     */
+    public CRTopologyTemplateResource(ComplianceRuleResource parent, TTopologyTemplate template) {
+        this.topologyTemplate = template;
+        this.complianceRuleRes = parent;
+    }
 
-	// @formatter:off
-	@GET
-	@ApiOperation(value = "Returns a JSON representation of the topology template. <br />" +
-		"X and Y coordinates are embedded as attributes. QName string with Namespace: <br />" +
-		"{@link org.eclipse.winery.repository.common.constants.Namespaces.TOSCA_WINERY_EXTENSIONS_NAMESPACE} <br />" +
-		"@return The JSON representation of the topology template <em>without</em> associated artifacts and without the parent service template")
-	@Produces(MediaType.APPLICATION_JSON)
-	// @formatter:on
-	public Response getComponentInstanceJSON() {
-		return Response.ok(this.topologyTemplate).build();
-	}
+    // @formatter:off
+    @GET
+    @ApiOperation(value = "Returns a JSON representation of the topology template. <br />" +
+        "X and Y coordinates are embedded as attributes. QName string with Namespace: <br />" +
+        "{@link org.eclipse.winery.repository.common.constants.Namespaces.TOSCA_WINERY_EXTENSIONS_NAMESPACE} <br />" +
+        "@return The JSON representation of the topology template <em>without</em> associated artifacts and without the parent service template")
+    @Produces(MediaType.APPLICATION_JSON)
+    // @formatter:on
+    public Response getComponentInstanceJSON() {
+        return Response.ok(this.topologyTemplate).build();
+    }
 
-	@PUT
-	@ApiOperation(value = "Replaces the topology by the information given in the XML")
-	@Consumes(MediaType.TEXT_XML)
-	public Response setModel(@Context UriInfo uriInfo, TTopologyTemplate topologyTemplate) {
-		return setTopologyAndPersist(uriInfo, topologyTemplate);
-	}
+    @PUT
+    @ApiOperation(value = "Replaces the topology by the information given in the XML")
+    @Consumes(MediaType.TEXT_XML)
+    public Response setModel(@Context UriInfo uriInfo, TTopologyTemplate topologyTemplate) {
+        return setTopologyAndPersist(uriInfo, topologyTemplate);
+    }
 
-	@PUT
-	@ApiOperation(value = "Replaces the topology by the information given in the XML")
-	@Consumes(MediaType.APPLICATION_JSON)
-	public Response setModelJson(@Context UriInfo uriInfo, TTopologyTemplate topologyTemplate) throws Exception {
-		ModelUtilities.patchAnyAttributes(topologyTemplate.getNodeTemplates());
-		setIdentifier(topologyTemplate);
-		return setTopologyAndPersist(uriInfo, topologyTemplate);
-	}
+    @PUT
+    @ApiOperation(value = "Replaces the topology by the information given in the XML")
+    @Consumes(MediaType.APPLICATION_JSON)
+    public Response setModelJson(@Context UriInfo uriInfo, TTopologyTemplate topologyTemplate) throws Exception {
+        ModelUtilities.patchAnyAttributes(topologyTemplate.getNodeTemplates());
+        return setTopologyAndPersist(uriInfo, topologyTemplate);
+    }
 
-	public Response setTopologyAndPersist(UriInfo uriInfo, TTopologyTemplate topologyTemplate) {
-		String[] split = uriInfo.getPath().split("/");
-		String lastSegment = split[split.length - 1];
+    public Response setTopologyAndPersist(UriInfo uriInfo, TTopologyTemplate topologyTemplate) {
+        String[] split = uriInfo.getPath().split("/");
+        String lastSegment = split[split.length - 1];
 
-		if (lastSegment.matches("identifier")) {
-			ComplianceRuleChecker checker = new ComplianceRuleChecker(topologyTemplate, this.complianceRuleRes.getCompliancerule().getRequiredStructure(), null);
-			if (!checker.isRuleValid()) {
-				return Response.status(Response.Status.BAD_REQUEST).build();
-			}
-			setIdentifier(topologyTemplate);
-		} else if (lastSegment.matches("requiredstructure")) {
-			ComplianceRuleChecker checker = new ComplianceRuleChecker(this.complianceRuleRes.getCompliancerule().getIdentifier(), topologyTemplate, null);
-			if (!checker.isRuleValid()) {
-				return Response.status(Response.Status.BAD_REQUEST).build();
-			}
-			setRequiredStructure(topologyTemplate);
-		}
-		return persist(this.complianceRuleRes);
-	}
+        if (lastSegment.matches("identifier")) {
+            ComplianceRuleChecker checker = new ComplianceRuleChecker(topologyTemplate, this.complianceRuleRes.getCompliancerule().getRequiredStructure(), null);
+            if (!checker.isRuleValid()) {
+                return Response.status(Response.Status.BAD_REQUEST).entity("Invalid Compliance Rule").build();
+            }
+            setIdentifier(topologyTemplate);
+        } else if (lastSegment.matches("requiredstructure")) {
+            ComplianceRuleChecker checker = new ComplianceRuleChecker(this.complianceRuleRes.getCompliancerule().getIdentifier(), topologyTemplate, null);
+            if (!checker.isRuleValid()) {
+                return Response.status(Response.Status.BAD_REQUEST).entity("Invalid Compliance Rule").build();
+            }
+            setRequiredStructure(topologyTemplate);
+        }
+        return persist(this.complianceRuleRes);
+    }
 
-	protected void setIdentifier(TTopologyTemplate topologyTemplate) {
-		this.complianceRuleRes.getCompliancerule().setIdentifier(topologyTemplate);
-	}
+    protected void setIdentifier(TTopologyTemplate topologyTemplate) {
+        this.complianceRuleRes.getCompliancerule().setIdentifier(topologyTemplate);
+    }
 
-	protected void setRequiredStructure(TTopologyTemplate topologyTemplate) {
-		this.complianceRuleRes.getCompliancerule().setRequiredStructure(topologyTemplate);
-	}
+    protected void setRequiredStructure(TTopologyTemplate topologyTemplate) {
+        this.complianceRuleRes.getCompliancerule().setRequiredStructure(topologyTemplate);
+    }
 
-	protected Response persist(ComplianceRuleResource complianceRuleRes) {
-		return RestUtils.persist(this.complianceRuleRes);
-	}
+    protected Response persist(ComplianceRuleResource complianceRuleRes) {
+        return RestUtils.persist(this.complianceRuleRes);
+    }
 
-	// @formatter:off
-	@GET
-	@ApiOperation(value = "<p>Returns an XML representation of the topology template." +
-		" X and Y coordinates are embedded as attributes. Namespace:" +
-		"{@link org.eclipse.winery.repository.common.constants.Namespaces.TOSCA_WINERY_EXTENSIONS_NAMESPACE} </p>" +
-		"<p>{@link org.eclipse.winery.repository.client.WineryRepositoryClient." +
-		"getTopologyTemplate(QName)} consumes this template</p>" +
-		"<p>@return The XML representation of the topology template <em>without</em>" +
-		"associated artifacts and without the parent service template </p>")
-	@Produces({MediaType.APPLICATION_XML, MediaType.TEXT_XML})
-	// @formatter:on
-	public Response getComponentInstanceXML() {
-		return RestUtils.getXML(TTopologyTemplate.class, this.topologyTemplate);
-	}
+    // @formatter:off
+    @GET
+    @ApiOperation(value = "<p>Returns an XML representation of the topology template." +
+        " X and Y coordinates are embedded as attributes. Namespace:" +
+        "{@link org.eclipse.winery.repository.common.constants.Namespaces.TOSCA_WINERY_EXTENSIONS_NAMESPACE} </p>" +
+        "<p>{@link org.eclipse.winery.repository.client.WineryRepositoryClient." +
+        "getTopologyTemplate(QName)} consumes this template</p>" +
+        "<p>@return The XML representation of the topology template <em>without</em>" +
+        "associated artifacts and without the parent service template </p>")
+    @Produces( {MediaType.APPLICATION_XML, MediaType.TEXT_XML})
+    // @formatter:on
+    public Response getComponentInstanceXML() {
+        return RestUtils.getXML(TTopologyTemplate.class, this.topologyTemplate);
+    }
 }
